Answer the review of the shrink-broadcast queue
Page the shrink over the chats instead of listing its candidates first. The
candidate query needed a DISTINCT over about a million stale dicks to save
visiting one chat in eight — nearly every chat has a neglected dick in it — so
it is replaced by a keyset walk of Chats' primary key, one page held at a time.

Report only what the alerts read, and only from the worker's tick. The gauge
over the finished rows grouped a few hundred thousand of them by state every
five seconds so that a graph could show what a single SQL query answers; the
Grafana Postgres datasource covers that, and it costs nothing when nobody is
looking. What is left is the two the alerts need, because vmalert reads
Prometheus and cannot query SQL. The shrink run and the cleaner no longer
report at all — the worker's tick subsumes both within a poll interval.

Settle what a log level means here: it follows what was lost, not whether the
code recovered. A failed shrink page is an error — those chats lost the day and
nothing retries it — while a failed metric publication stays a warning, since
the next tick replaces the sample and the database problem behind it has
already been reported by the run that hit it. The per-page line carries the
error, so the summary at the end is no longer a second error saying less.

Share the back-off between the two queues, and drop the trait that let a shrink
page render from two row types when only one is left.

Give Count an Add to go with its AddAssign. Not Sub: saturating at zero would
turn a difference that was never meant to go negative into a plausible 0.

Move both queues' tests under repo/test/, where every other repository's live.

CLAUDE.md claimed Gauge::set and Histogram::observe take the conversion traits
so that callers hand over domain values. They don't, and they can't:
domain_types implements those traits only for the pairs that lose something, on
the principle that an exact conversion says From — so there is no
SaturatingInto<i64> for i64, and a caller holding a timestamp could not satisfy
such a bound. The callers were right and the document was wrong; it now says so.
