+++ .sqlx/query-33c5bc97a1111bb38d68d7445d97e3ebdfe0f2462b281ac16125261ccc564ec8.json
@@ -1,0 +1,29 @@
+{
+  "db_name": "PostgreSQL",
+  "query": "SELECT id AS \"id: InternalChatId\" FROM Chats\n                WHERE id > $1 ORDER BY id LIMIT $2",
+  "describe": {
+    "columns": [
+      {
+        "ordinal": 0,
+        "name": "id: InternalChatId",
+        "type_info": "Int8",
+        "origin": {
+          "Table": {
+            "table": "chats",
+            "name": "id"
+          }
+        }
+      }
+    ],
+    "parameters": {
+      "Left": [
+        "Int8",
+        "Int8"
+      ]
+    },
+    "nullable": [
+      false
+    ]
+  },
+  "hash": "33c5bc97a1111bb38d68d7445d97e3ebdfe0f2462b281ac16125261ccc564ec8"
+}
