--- .sqlx/query-eefbd7ed6be814919face127928a1353e4b672981b39b3552c507efb47b48e15.json
@@ -1,28 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "SELECT DISTINCT chat_id AS \"chat_id: InternalChatId\" FROM Dicks\n                WHERE length > 0 AND updated_at <= current_timestamp - make_interval(days => $1::bigint::int)",
-  "describe": {
-    "columns": [
-      {
-        "ordinal": 0,
-        "name": "chat_id: InternalChatId",
-        "type_info": "Int8",
-        "origin": {
-          "Table": {
-            "table": "dicks",
-            "name": "chat_id"
-          }
-        }
-      }
-    ],
-    "parameters": {
-      "Left": [
-        "Int8"
-      ]
-    },
-    "nullable": [
-      false
-    ]
-  },
-  "hash": "eefbd7ed6be814919face127928a1353e4b672981b39b3552c507efb47b48e15"
-}
