--- .sqlx/query-fa12c6a161068cc2f0360936acb7939d4f2d8c64fc8da71ccf0c0c98a0e19f42.json
@@ -1,12 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "INSERT INTO Scheduled_Message_Deletions (message_kind, message_group, lang_code, fire_after)\n                VALUES ('reply', 'notice', 'en', current_timestamp - interval '1 minute')",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Left": []
-    },
-    "nullable": []
-  },
-  "hash": "fa12c6a161068cc2f0360936acb7939d4f2d8c64fc8da71ccf0c0c98a0e19f42"
-}
